move statement test, rename packages for consistency
