+++ src/main/java/com/abc/Account/CheckingAccount.java
@@ -1,8 +1,8 @@
-package com.abc.Accounts;
+package com.abc.Account;
 
 public class CheckingAccount extends Account {
 
-    private static final String ACCOUNT_NAME = "Checking Accounts";
+    private static final String ACCOUNT_NAME = "Checking Account";
 
     public String getName(){
         return ACCOUNT_NAME;
