+++ src/main/java/com/abc/Account/MaxiSavingsAccount.java
@@ -1,8 +1,8 @@
-package com.abc.Accounts;
+package com.abc.Account;
 
 public class MaxiSavingsAccount extends Account {
 
-    private static final String ACCOUNT_NAME = "Maxi-Savings Accounts";
+    private static final String ACCOUNT_NAME = "Maxi-Savings Account";
 
     public String getName() {
         return ACCOUNT_NAME;
