+++ src/test/java/com/abc/BankTest.java
@@ -1,9 +1,9 @@
 package com.abc;
 
-import com.abc.Accounts.Account;
-import com.abc.Accounts.CheckingAccount;
-import com.abc.Accounts.MaxiSavingsAccount;
-import com.abc.Accounts.SavingsAccount;
+import com.abc.Account.Account;
+import com.abc.Account.CheckingAccount;
+import com.abc.Account.MaxiSavingsAccount;
+import com.abc.Account.SavingsAccount;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
